requirements.txt
# FastAPI framework
fastapi==0.104.1
uvicorn==0.24.0

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
cryptography==41.0.7
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6

# Validation
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator==2.1.0

# Environment variables
python-dotenv==1.0.0

# HTTP client
httpx==0.25.2
tenacity==8.2.3

# Google OAuth
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Email - Brevo (Sendinblue)
sib-api-v3-sdk==7.6.0

# Rate Limiting
fastapi-limiter==0.1.6
redis==5.0.1